[SC] fix: Stats for max level (#4)
* small changes

(cherry picked from commit 18fc2aaa655ad74bd932397da42854147981a217)

* fixed custom player stats

(cherry picked from commit a3b5e269b93651d76e4246ca93957a1499d55a27)

* Gozzim/Dokito - Changed Talent Points calculation

(cherry picked from commit e16eeb200a028d757520733d4038072342ca222e)

* feat: various stat changes

+++ src/server/game/Globals/ObjectMgr.h
@@ -523,8 +523,8 @@ struct PetLevelInfo
     PetLevelInfo() { for (unsigned short & stat : stats) stat = 0; }
 
     uint16 stats[MAX_STATS];
-    uint16 health{0};
-    uint16 mana{0};
+    uint32 health{0};
+    uint32 mana{0};
     uint32 armor{0};
     uint16 min_dmg{0};
     uint16 max_dmg{0};
